+++ CharacterOptionsPlus/changelog.md
@@ -17,6 +17,7 @@
 
 * New Bugfixes:
     * Cone of Cold is now on the Witch spell list
+    * Evasion is now available as an Advanced Slayer Talent
 
 * New Optional Changes:
     * Cone of Cold replaces Cold Ice Strike as a bonus spell granted by the Winter Patron